refactor(pipeline): extract destroy macro from monoliths into atomic primitives (#192)
Replace duplicated 8-step destroy logic in job_worker.py and esacp.py
with macro/destroy.py composing 8 single-task IoC primitives. Delete
superseded destroy_helpers.py. E2E tested: dev01 fully destroyed via CLI.

fixes #192
