+++ tools/generate_inventory.py
