+++ tools/size_baselines.json
@@ -5,24 +5,33 @@
   "tools/job_worker.py": 305,
   "tools/pipeline/__init__.py": 0,
   "tools/pipeline/macro/__init__.py": 0,
+  "tools/pipeline/macro/destroy.py": 75,
   "tools/pipeline/macro/provision.py": 86,
   "tools/pipeline/macro/provision_generic.py": 92,
   "tools/pipeline/macro/refresh.py": 41,
   "tools/pipeline/orchestration/__init__.py": 0,
+  "tools/pipeline/orchestration/ansible_wg_update.py": 36,
   "tools/pipeline/orchestration/build_template.py": 77,
   "tools/pipeline/orchestration/build_vm.py": 75,
   "tools/pipeline/orchestration/build_vm_create.py": 70,
   "tools/pipeline/orchestration/build_vm_poll.py": 68,
   "tools/pipeline/orchestration/build_vm_seed.py": 46,
+  "tools/pipeline/orchestration/cloud_init_cleanup.py": 17,
   "tools/pipeline/orchestration/destroy_vm.py": 57,
+  "tools/pipeline/orchestration/group_vars_remove.py": 25,
+  "tools/pipeline/orchestration/hosts_map_remove.py": 26,
   "tools/pipeline/orchestration/hypervisor_helpers.py": 67,
+  "tools/pipeline/orchestration/inventory_regen.py": 20,
   "tools/pipeline/orchestration/load_host_config.py": 18,
   "tools/pipeline/orchestration/memory_guard.py": 77,
   "tools/pipeline/orchestration/observability_creds.py": 73,
+  "tools/pipeline/orchestration/sops_key_remove.py": 73,
   "tools/pipeline/orchestration/verify_build_vm.py": 80,
   "tools/pipeline/orchestration/verify_vpn.py": 77,
   "tools/pipeline/orchestration/virsh.py": 15,
   "tools/pipeline/orchestration/vm_power.py": 46,
+  "tools/pipeline/orchestration/wg_peer_remove.py": 28,
+  "tools/pipeline/orchestration/wg_pubkey.py": 32,
   "tools/pipeline/runner.py": 27,
   "tools/pipeline/stages/__init__.py": 0,
   "tools/pipeline/stages/common/__init__.py": 0,
